requirements.txt
mysql-connector-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
configparser